Add aggregateRating to output.
This patch fixes a problem whereby Google Structured Data Testing Tool complains
"Either “offers”, “review”, or “aggregateRating” should be specified".
I figure it's technically correct that, for a product in a review post,
the aggregate rating would be whatever the rating metadata says, and the
number of applicable reviews is 1 - this seemed like the most
appropriate way of satisfying the checker.